+++ packages/anov-3d-utils/src/index.ts
@@ -1,10 +1,10 @@
-import { Scene } from '@anov/3d'
+import { SceneControl } from '@anov/3d'
 import { mouseDraweLine } from './core/mouseDrawe'
 import { createSnow } from './environment/snow'
 import { createRain } from './environment/rain'
 
 export {
-  Scene, // todo remove
+  SceneControl, // todo remove
   mouseDraweLine,
   createRain,
   createSnow,