+++ playground/example/example-03.ts
@@ -1,4 +1,4 @@
-import { ModelLoader, Scene } from '@anov/3d'
+import { ModelLoader, SceneControl as Scene } from '@anov/3d'
 
 /**
  * example-03