+++ playground/example/example-12.ts
@@ -1,4 +1,4 @@
-import { Mesh, Scene, BoxGeometry, MeshBasicMaterial, createTransformControls, TransformControls } from '../../packages/anov-3d/src/index'
+import { Mesh, SceneControl as Scene, BoxGeometry, MeshBasicMaterial, createTransformControls, TransformControls } from '../../packages/anov-3d/src/index'
 
 
 /**