+++ playground/example/example-13.ts
@@ -1,4 +1,4 @@
-import { Color, Vector3, Scene, createTransformControls, createControlLine } from '../../packages/anov-3d/src/index'
+import { Color, Vector3, SceneControl as Scene, createTransformControls, createControlLine } from '../../packages/anov-3d/src/index'
 
 
 /**