+++ packages/anov-3d/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @anov/3d
 
+## 0.0.3
+
+### Patch Changes
+
+- fix: adjust api Scene->SceneControl
+
 ## 0.0.2
 
 ### Patch Changes