+++ playground/CHANGELOG.md
@@ -5,6 +5,14 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @anov/3d-utils@0.0.3
+  - @anov/3d@0.0.3
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
   - @anov/3d@0.0.2
 
 ## 1.0.1